requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
requests
peewee
PyMySQL==1.0.2
python-dotenv==0.21.0
flask-cors

Flask-JWT==0.3.2

python-json-logger
pylint
pylint-runner
flask-socketio
flask-sock

neo4j
passlib
apscheduler
flask_swagger_ui
